+++ mithril-signer/src/configuration.rs
@@ -108,6 +108,11 @@ pub struct Configuration {
     /// [network_security_parameter][Self::network_security_parameter] blocks after each import
     /// `[default: true]`.
     pub enable_transaction_pruning: bool,
+
+    /// Chunk size for importing transactions, combined with transaction pruning it reduces the
+    /// storage footprint of the signer by reducing the number of transactions stored on disk
+    /// at any given time.
+    pub transactions_import_block_chunk_size: BlockNumber,
 }
 
 impl Configuration {
@@ -143,6 +148,7 @@ impl Configuration {
             metrics_server_port: 9090,
             allow_unparsable_block: false,
             enable_transaction_pruning: false,
+            transactions_import_block_chunk_size: 1000,
         }
     }
 
@@ -212,6 +218,9 @@ pub struct DefaultConfiguration {
 
     /// Preload security parameter
     pub preload_security_parameter: BlockNumber,
+
+    /// Chunk size for importing transactions
+    pub transactions_import_block_chunk_size: BlockNumber,
 }
 
 impl DefaultConfiguration {
@@ -229,6 +238,7 @@ impl Default for DefaultConfiguration {
             network_security_parameter: 2160, // 2160 is the mainnet value
             preload_security_parameter: 3000,
             enable_transaction_pruning: true,
+            transactions_import_block_chunk_size: 1500,
         }
     }
 }
@@ -273,6 +283,11 @@ impl Source for DefaultConfiguration {
         result.insert(
             "enable_transaction_pruning".to_string(),
             into_value(myself.enable_transaction_pruning),
+        );
+
+        result.insert(
+            "transactions_import_block_chunk_size".to_string(),
+            into_value(myself.transactions_import_block_chunk_size),
         );
 
         Ok(result)